Corona: Modernize the Model class

+++ src/Lunr/Corona/Model.php
@@ -10,6 +10,7 @@
 namespace Lunr\Corona;
 
 use Psr\Cache\CacheItemPoolInterface;
+use RuntimeException;
 
 /**
  * Base Model class
@@ -21,7 +22,7 @@ class Model
      * Shared instance of the cache Pool class or null.
      * @var CacheItemPoolInterface|null
      */
-    protected $cache;
+    protected readonly ?CacheItemPoolInterface $cache;
 
     /**
      * Constructor.
@@ -38,7 +39,21 @@ class Model
      */
     public function __destruct()
     {
-        unset($this->cache);
+        // no-op
+    }
+
+    /**
+     * Get cache item.
+     *
+     * @deprecated Use getFromCache() instead
+     *
+     * @param string $key Unique identifier for the data.
+     *
+     * @return mixed Cached data
+     */
+    protected function get_from_cache(string $key)
+    {
+        return $this->getFromCache($key);
     }
 
     /**
@@ -48,9 +63,30 @@ class Model
      *
      * @return mixed Cached data
      */
-    protected function get_from_cache(string $key)
+    protected function getFromCache(string $key): mixed
     {
+        if (!isset($this->cache))
+        {
+            throw new RuntimeException('No Cachepool initialized!');
+        }
+
         return $this->cache->getItem($key)->get();
+    }
+
+    /**
+     * Set cache item.
+     *
+     * @deprecated Use setInCache() instead
+     *
+     * @param string $key   Unique identifier for the data.
+     * @param mixed  $value Data to cache.
+     * @param int    $ttl   Time To Live for cache item in seconds.
+     *
+     * @return bool True if successful, FALSE when value is NULL.
+     */
+    protected function set_in_cache(string $key, $value, int $ttl = 600): bool
+    {
+        return $this->setInCache($key, $value, $ttl);
     }
 
     /**
@@ -62,11 +98,16 @@ class Model
      *
      * @return bool True if successful, FALSE when value is NULL.
      */
-    protected function set_in_cache(string $key, $value, int $ttl = 600): bool
+    protected function setInCache(string $key, mixed $value, int $ttl = 600): bool
     {
         if ($value === NULL)
         {
             return FALSE;
+        }
+
+        if (!isset($this->cache))
+        {
+            throw new RuntimeException('No Cachepool initialized!');
         }
 
         $item = $this->cache->getItem($key);
@@ -82,13 +123,29 @@ class Model
     /**
      * Get an item from the cache or seed the cache if no item is present.
      *
-     * @param string   $id       ID of the item
-     * @param callable $callable The callable to seed the cache with
-     * @param array    $args     Arguments to the callback
+     * @deprecated Use cacheIfNeeded() instead
+     *
+     * @param string                  $id       ID of the item
+     * @param callable                $callable The callable to seed the cache with
+     * @param array<array-key, mixed> $args     Arguments to the callback
      *
      * @return mixed Cache or Callable return
      */
     protected function cache_if_needed(string $id, callable $callable, array $args = [])
+    {
+        return $this->cacheIfNeeded($id, $callable, $args);
+    }
+
+    /**
+     * Get an item from the cache or seed the cache if no item is present.
+     *
+     * @param string                  $id       ID of the item
+     * @param callable                $callable The callable to seed the cache with
+     * @param array<array-key, mixed> $args     Arguments to the callback
+     *
+     * @return mixed Cache or Callable return
+     */
+    protected function cacheIfNeeded(string $id, callable $callable, array $args = []): mixed
     {
         if ($this->cache === NULL)
         {
@@ -102,9 +159,23 @@ class Model
         }
 
         $value = call_user_func_array($callable, $args);
-        $this->set_in_cache($id, $value);
+        $this->setInCache($id, $value);
 
         return $value;
+    }
+
+    /**
+     * Delete cache item.
+     *
+     * @deprecated Use deleteFromCache() instead
+     *
+     * @param string $key Unique identifier for the data.
+     *
+     * @return bool True if the item was successfully removed. False if there was an error.
+     */
+    protected function delete_from_cache(string $key): bool
+    {
+        return $this->deleteFromCache($key);
     }
 
     /**
@@ -114,8 +185,13 @@ class Model
      *
      * @return bool True if the item was successfully removed. False if there was an error.
      */
-    protected function delete_from_cache(string $key): bool
+    protected function deleteFromCache(string $key): bool
     {
+        if (!isset($this->cache))
+        {
+            throw new RuntimeException('No Cachepool initialized!');
+        }
+
         return $this->cache->deleteItem($key);
     }
